# Increase chunk size for performance