Increasing chunk size from 1MB to 4MB.
Most of the time are spent on sending POST request to server.
When sending an 1.14GB file:
```text
`getStfFile`: 5210ms 
`fetch /api/upload`: 34300ms
```

I think it would be better to send large payload with fewer requests would help in this case. 4MB is not big at all as of now.
Also, reading the files in big chunks would also help.
After changing the chunk size, even though it's local, same host, dev server:
```text
`getStfFile`: 3179ms 
`fetch /api/upload`: 24758ms
```

Big improvement.